requirements.txt
pip==20.1.1
pycryptodomex==3.10.1
setuptools==47.1.0
